Add placeholder for @gensx/cli.
