+++ .release-please-manifest.json
@@ -1,9 +1,10 @@
 {
-  "packages/gensx-core": "0.1.0",
-  "packages/gensx-openai": "0.1.10",
   "packages/create-gensx": "0.1.11",
-  "packages/gensx-cli": "0.1.5",
-  "packages/gensx-vercel-ai-sdk": "0.1.3",
   "packages/gensx-anthropic": "0.1.2",
-  "packages/gensx-mcp": "0.1.3"
+  "packages/gensx-cli": "0.1.5",
+  "packages/gensx-core": "0.1.0",
+  "packages/gensx-mcp": "0.1.3",
+  "packages/gensx-openai": "0.1.10",
+  "packages/gensx-vercel-ai-sdk": "0.1.3",
+  "packages/gensx": "0.3.0"
 }
