+++ packages/gensx-cli/README.md
@@ -1,37 +1,3 @@
-# GenSX CLI
+# @gensx/cli
 
-## Installation
-
-```bash
-npm install -g gensx
-
-# or
-
-pnpm install -g gensx
-
-# or
-
-yarn global add gensx
-
-# or
-
-brew install gensx-inc/gensx/gensx
-```
-
-Or use without installing globally:
-
-```bash
-npx gensx
-```
-
-## Usage
-
-### Login
-
-Login to GenSX to get started:
-
-```bash
-gensx login
-```
-
-Now you will have an API key saved in your config file, and this will be used by default for any commands that require authentication, or when running a workflow locally.
+This is deprecated, please use the `gensx` package instead.
