+++ packages/gensx-cli/package.json
@@ -1,64 +1,33 @@
 {
-  "name": "gensx",
-  "version": "0.1.5",
-  "description": "`GenSX command line tools",
+  "name": "@gensx/cli",
+  "version": "0.1.11",
+  "type": "module",
+  "license": "Apache-2.0",
+  "engines": {
+    "node": ">=18.0.0",
+    "pnpm": ">=9.0.0"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/gensx-inc/gensx.git",
     "directory": "packages/gensx-cli"
   },
-  "bin": {
-    "gensx": "./dist/run.js"
-  },
+  "bin": "./dist/index.js",
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
-  "directories": {
-    "lib": "src",
-    "bin": "bin"
-  },
   "files": [
-    "dist",
-    "bin"
+    "dist"
   ],
   "scripts": {
-    "build": "tsc -p tsconfig.build.json && cp -r src/templates dist/",
+    "build": "tsc -p tsconfig.build.json",
     "clean": "rm -rf dist",
-    "compile": "tsc",
+    "test": "vitest run --coverage",
+    "test:watch": "vitest",
     "lint": "eslint .",
-    "lint:fix": "eslint . --fix",
-    "start": "tsx ./bin/run.ts",
-    "start:node": "node ./bin/run",
-    "package": "pkg -t $TARGET --compress GZip --out-path dist/pkg ./dist/run.cjs",
-    "package:macos": "TARGET=node22-macos-arm64 pnpm package"
+    "lint:fix": "eslint . --fix"
   },
-  "keywords": [
-    "typescript",
-    "gensx",
-    "cli"
-  ],
-  "author": "GenSX Team",
-  "license": "Apache-2.0",
+  "dependencies": {},
   "devDependencies": {
-    "@types/ini": "^4.1.1",
-    "@types/node": "catalog:",
-    "@vitest/coverage-istanbul": "catalog:",
-    "@yao-pkg/pkg": "^6.3.0",
-    "vitest": "catalog:"
-  },
-  "dependencies": {
-    "commander": "^11.0.0",
-    "consola": "^3.2.3",
-    "enquirer": "^2.4.1",
-    "ini": "^5.0.0",
-    "open": "^10.0.3",
-    "ora": "^8.2.0",
-    "picocolors": "^1.0.1"
-  },
-  "type": "module",
-  "publishConfig": {
-    "access": "public"
-  },
-  "pkg": {
-    "scripts": "dist/run.cjs"
+    "@types/node": "catalog:"
   }
 }
