+++ packages/gensx/.npmignore
