+++ packages/gensx/.npmrc
