+++ packages/gensx/CHANGELOG.md
@@ -1,0 +1,65 @@
+# Changelog
+
+## [0.1.5](https://github.com/gensx-inc/gensx/compare/gensx-cli-v0.1.4...gensx-cli-v0.1.5) (2025-03-04)
+
+
+### 🐛 Bug Fixes
+
+* Fix build ([#399](https://github.com/gensx-inc/gensx/issues/399)) ([fa26f63](https://github.com/gensx-inc/gensx/commit/fa26f63ac688f0be423a9a6ce6585b7600bb6cb1))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * create-gensx bumped to 0.1.11
+
+## [0.1.4](https://github.com/gensx-inc/gensx/compare/gensx-cli-v0.1.3...gensx-cli-v0.1.4) (2025-03-03)
+
+
+### ✨ New Features
+
+* @gensx/mcp ([#358](https://github.com/gensx-inc/gensx/issues/358)) ([37d316e](https://github.com/gensx-inc/gensx/commit/37d316e0819e3ae9ffc74f568170ff84fff4da3c))
+* Don't minify packages ([#371](https://github.com/gensx-inc/gensx/issues/371)) ([2761ab8](https://github.com/gensx-inc/gensx/commit/2761ab862319858bd0447c2d006903f91d9a9b79))
+
+## [0.1.3](https://github.com/gensx-inc/gensx/compare/gensx-cli-v0.1.2...gensx-cli-v0.1.3) (2025-02-22)
+
+
+### ✨ New Features
+
+* Prompt for login when calling new the first time. ([#307](https://github.com/gensx-inc/gensx/issues/307)) ([271dca4](https://github.com/gensx-inc/gensx/commit/271dca4a35be2403f1065dfbb9948dbbfa14920c))
+
+## [0.1.2](https://github.com/gensx-inc/gensx/compare/gensx-cli-v0.1.1...gensx-cli-v0.1.2) (2025-02-20)
+
+
+### ✨ New Features
+
+* Colorize login command. ([#303](https://github.com/gensx-inc/gensx/issues/303)) ([36af168](https://github.com/gensx-inc/gensx/commit/36af168cab8faebad5e11d40516075ac9b2ed84d))
+* Print URL to link to execution. ([#293](https://github.com/gensx-inc/gensx/issues/293)) ([3f6898e](https://github.com/gensx-inc/gensx/commit/3f6898e5cc2a02e53286f939528f6fa499b52238))
+
+
+### 🐛 Bug Fixes
+
+* Don't use global crypto. ([#298](https://github.com/gensx-inc/gensx/issues/298)) ([8ce57d8](https://github.com/gensx-inc/gensx/commit/8ce57d863185a88cb14a7f72ac70f7894b25164e))
+
+## [0.1.1](https://github.com/gensx-inc/gensx/compare/gensx-cli-v0.1.0...gensx-cli-v0.1.1) (2025-02-19)
+
+
+### 🐛 Bug Fixes
+
+* Add hashbang for using npx gensx ([#286](https://github.com/gensx-inc/gensx/issues/286)) ([5aae4ec](https://github.com/gensx-inc/gensx/commit/5aae4ecd6786e45be77b64a7b7d6320acbe63775))
+
+## 0.1.0 (2025-02-19)
+
+
+### ✨ New Features
+
+* CLI Device auth flow ([#212](https://github.com/gensx-inc/gensx/issues/212)) ([094b98e](https://github.com/gensx-inc/gensx/commit/094b98e12ef4239e8b04c176a14f19f5e891f5a1))
+* CLI new command ([#280](https://github.com/gensx-inc/gensx/issues/280)) ([5a8743c](https://github.com/gensx-inc/gensx/commit/5a8743c0df6aa188d45239cc96169ef0671c146e))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * create-gensx bumped to 0.1.8
