+++ packages/gensx/CONTRIBUTING.md
