+++ packages/gensx/eslint.config.mjs
@@ -1,13 +1,16 @@
 import rootConfig from "../../eslint.config.mjs";
 
 export default [
+  {
+    ignores: ["**/coverage/**"],
+  },
   ...rootConfig,
   {
     files: ["**/*.{js,mjs,cjs,jsx,ts,tsx}"],
     languageOptions: {
       parserOptions: {
-        tsconfigRootDir: import.meta.dirname,
         project: "./tsconfig.json",
+        tsconfigRootDir: import.meta.dirname,
       },
     },
   },
