+++ packages/gensx/src/commands/new.ts
