+++ packages/gensx/src/templates/projects/typescript/README.md.template
