+++ packages/gensx/src/templates/projects/typescript/nodemon.json.template
