+++ packages/gensx/src/templates/projects/typescript/package.json.template
