+++ packages/gensx/src/templates/projects/typescript/src/index.tsx.template
