+++ packages/gensx/src/templates/projects/typescript/template.json
