+++ packages/gensx/src/templates/projects/typescript/tsconfig.json.template
