+++ packages/gensx/src/utils/config.ts
