+++ packages/gensx/src/utils/terminal.ts
